// src/types/articulo.ts
export interface Articulo {
  Codigo: string;
  Descripcion: string;
  ProveedorCodigo: string;
  RubroCodigo: string;
  PrecioCosto: number;
  Lista1: number;
  Lista2: number;
  Lista3: number;
  Lista4: number;
  Lista5: number;
  Existencia: number;
  StockMinimo: number;
  StockMaximo: number;
}

// export interface ProductReportItem {
//   FacturaNumero: string;
//   ProductoCodigo: string;
//   ProductoDescripcion: string;
//   ProveedorCodigo: string;
//   Cantidad: number;
//   ClienteDescripcion: string;
//   ClienteCUIT: string;
// }